refactor: Service 인터페이스/구현체 네이밍 컨벤션 수정
인터페이스와 구현체의 이름이 역전되어 있던 문제 해결
- LectureServiceImpl(interface) → LectureService
- LectureService(impl) → LectureServiceImpl
